Switch to new GQA
