+++ onnxscript/rewriter/ort_fusions/gqa.py
@@ -2,149 +2,232 @@
 # Licensed under the MIT License.
 from __future__ import annotations
 
+from typing import Sequence, Union
+
 import onnxscript.ir as ir
-from onnxscript.optimizer import remove_unused_nodes
 from onnxscript.rewriter import pattern
+
+"""
+GroupQueryAttention: This generalizes MHA by allowing the number of heads to be different
+for query and key/value.
+
+We use the following abbreviations for the dimensions:
+B: Batch size
+S: Sequence length (for current query/key/value)
+D: input embedding dimension
+Dkv: key/value hidden size
+H: number of heads (must be an integral multiple of Hkv)
+Hkv: number of heads for key/value
+Dh: head size or embedding dimension per head (usually, D = H * Dh)
+Skv: key/value sequence length
+St: total sequence length
+
+In the sequel, the suffix "_BHSDh" indicates that the tensor has the shape (B, H, S, Dh).
+The suffix "BH_Skv_Dh" indicates that the tensor has the shape (B*H, Skv, Dh).
+"""
+
+Dim = Union[int, ir.SymbolicDim]
+
+
+def _check_shape(bindings: dict[str, Dim], val: ir.Value, shape: Sequence[str]) -> bool:
+    if val.shape is None:
+        return False
+    if val.shape.rank() != len(shape):
+        return False
+    for actual, expected in zip(val.shape, shape):
+        if expected not in bindings:
+            bindings[expected] = actual  # type: ignore[assignment]
+        elif actual != bindings[expected]:
+            return False
+    return True
 
 
 class GroupQueryAttention(pattern.RewriteRuleClassBase):
-    def __init__(self, name: str, *, use_2d_matmul: bool):
-        super().__init__(name, remove_nodes=False)
-        self._use_2d_matmul = use_2d_matmul
-
-    def _compute_packed_QKV(self, op, input, weight):
-        if self._use_2d_matmul:
-            # Convert batched input of shape (B, S, D) to 2D input (B*S, D)
-            input = op.Reshape(input, _allow_other_inputs=True)
-        projected = op.MatMul(input, weight)
-        if self._use_2d_matmul:
-            # Convert 2D output back to batched output of shape (B, S, D)
-            projected = op.Reshape(projected, _allow_other_inputs=True)
-        # Split combined QKV into Q, K, and V
-        query_3d = op.Slice(projected, _allow_other_inputs=True)
-        key_3d = op.Slice(projected, _allow_other_inputs=True)
-        value_3d = op.Slice(projected, _allow_other_inputs=True)
-        # Reshape from (B, S, D) to (B, S, H, D/H)
-        query_4d = op.Reshape(
-            query_3d,
-            _allow_other_inputs=True,
-            _allow_other_attributes=True,
-            _outputs=["query_mm_reshaped"],
-        )
-        # Transpose from (B, S, H, D/H) to (B, H, S, D/H)
-        query = op.Transpose(query_4d, perm=[0, 2, 1, 3])
-        key_4d = op.Reshape(
-            key_3d,
-            _allow_other_inputs=True,
-            _allow_other_attributes=True,
-            _outputs=["key_mm_reshaped"],
-        )
-        key = op.Transpose(key_4d, perm=[0, 2, 1, 3])
-        value_4d = op.Reshape(
-            value_3d,
-            _allow_other_inputs=True,
-            _allow_other_attributes=True,
-            _outputs=["value_mm_reshaped"],
-        )
-        value = op.Transpose(value_4d, perm=[0, 2, 1, 3])
-
-        return query, key, value
+    def __init__(self):
+        super().__init__("GQA")
 
     def pattern(
         self,
         op,
-        input,
-        qkv_weight,
+        query_BSD,
+        key_BSDkv,
+        value_BSDkv,
         mask,
-        cos,
-        sin,
         past_key,
         past_value,
         position_ids,
+        cos,
+        sin,
     ):
-        query, key, value = self._compute_packed_QKV(op, input, qkv_weight)
-
-        query_rope = op.RotaryEmbedding(query, position_ids, cos, sin, _domain="com.microsoft")
-
-        key_rope = op.RotaryEmbedding(key, position_ids, cos, sin, _domain="com.microsoft")
-        present_key = op.Concat(past_key, key_rope, axis=-2)
-        # Transpose last two axes of present_key to compute dot-product via matmul.
-        present_key = op.Transpose(present_key, perm=[0, 1, 3, 2])
-
-        present_value = op.Concat(past_value, value, axis=-2)
-
-        attention = op.SDPA(
-            query_rope, present_key, present_value, mask, _domain="ai.onnxruntime.fusion"
+        # Reshape query from (B, S, D) to (B, S, H, D/H)
+        query_BSHDh = op.Reshape(
+            query_BSD,
+            _allow_other_inputs=True,
+            _allow_other_attributes=True,
+            _outputs=["query_BSHDh"],
         )
-        # Transpose back to (B, S, H, D/H)
-        attention_transposed = op.Transpose(attention, perm=[0, 2, 1, 3])
+        # Transpose from (B, S, H, D/H) to (B, H, S, D/H)
+        query_BHSDh = op.Transpose(query_BSHDh, perm=[0, 2, 1, 3])
+
+        # Reshape key from (B, S, Dkv) to (B, S, Hkv, D/H)
+        key_BSHkvDh = op.Reshape(
+            key_BSDkv,
+            _allow_other_inputs=True,
+            _allow_other_attributes=True,
+            _outputs=["key_BSHkvDh"],
+        )
+        # Transpose from (B, S, Hkv, D/H) to (B, Hkv, S, D/H)
+        key_BHkvSDh = op.Transpose(key_BSHkvDh, perm=[0, 2, 1, 3])
+
+        # Reshape value from (B, S, Dkv) to (B, S, Hkv, D/H)
+        value_BSHkvDh = op.Reshape(
+            value_BSDkv,
+            _allow_other_inputs=True,
+            _allow_other_attributes=True,
+            _outputs=["value_BSHkvDh"],
+        )
+        # Transpose from (B, S, Hkv, D/H) to (B, Hkv, S, D/H)
+        value_BHkvSDh = op.Transpose(value_BSHkvDh, perm=[0, 2, 1, 3])
+
+        position_ids_q = op.Unsqueeze(position_ids, [0])
+        position_ids_k = op.Unsqueeze(position_ids, [0])
+
+        query_BHSDh_rope = op.RotaryEmbedding(
+            query_BHSDh, position_ids_q, cos, sin, _domain="com.microsoft"
+        )
+        key_BHkvSDh_rope = op.RotaryEmbedding(
+            key_BHkvSDh, position_ids_k, cos, sin, _domain="com.microsoft"
+        )
+
+        # Concatenate past_key cache and current key, expand across heads
+        # that share key/value and transpose to enable dot-product attention computation.
+
+        key_seq_BHkvSDh = op.Concat(past_key, key_BHkvSDh_rope, axis=-2)
+        key_seq_BHkv1SDh = op.Unsqueeze(key_seq_BHkvSDh, 2)
+        key_seq_BHkvGSDh = op.Expand(key_seq_BHkv1SDh, _allow_other_inputs=True)
+        key_seq_BHSkvDh = op.Reshape(
+            key_seq_BHkvGSDh, _allow_other_inputs=True, _outputs=["key_seq_BHSkvDh"]
+        )
+        key_seq_BHDhSkv = op.Transpose(
+            key_seq_BHSkvDh, _allow_other_inputs=True, _outputs=["key_seq_BHDhSkv"]
+        )
+
+        # Concatenate past_value cache and current value, expand across heads
+        # that share key/value.
+        value_seq_BHkvSDh = op.Concat(past_value, value_BHkvSDh, axis=-2)
+        value_seq_BHkv1SDh = op.Unsqueeze(value_seq_BHkvSDh, 2)
+        value_seq_BHkvGSDh = op.Expand(value_seq_BHkv1SDh, _allow_other_inputs=True)
+        value_seq_BHSkvDh = op.Reshape(
+            value_seq_BHkvGSDh, _allow_other_inputs=True, _outputs=["value_seq_BHSkvDh"]
+        )
+
+        attention_BHSDh = op.SDPA(
+            query_BHSDh_rope,
+            key_seq_BHDhSkv,
+            value_seq_BHSkvDh,
+            mask,
+            _domain="ai.onnxruntime.fusion",
+        )
+
+        # Transpose attention back to (B, S, H, D/H)
+        attention_BSHDh = op.Transpose(attention_BHSDh, perm=[0, 2, 1, 3])
         # Reshape back to (B, S, D)
-        attention_reshaped = op.Reshape(
-            attention_transposed, _allow_other_inputs=True, _outputs=["attention_reshaped"]
+        attention_BSD = op.Reshape(
+            attention_BSHDh, _allow_other_inputs=True, _outputs=["attention_BSD"]
         )
-        return attention_reshaped, present_key, present_value
+        return attention_BSD, key_seq_BHkvSDh, value_seq_BHkvSDh
 
     def check(
         self,
         op,
-        # query_mm_reshaped,
-        # key_mm_reshaped,
-        # value_mm_reshaped,
-        # key_reshaped,
-        # key_transposed,
-        # attention_reshaped,
+        query_BSD,
+        key_BSDkv,
+        value_BSDkv,
+        mask,
+        past_key,
+        past_value,
+        # query_BSHDh,
+        # key_BSHkvDh,
+        # value_BSHkvDh,
         **_,
     ):
-        # bindings: dict[str, int] = {}
-        # status = (
-        #     _check_shape(bindings, query_mm_reshaped, ["B", "S", "H", "d_h"])
-        #     and _check_shape(bindings, key_mm_reshaped, ["B", "S", "H", "d_h"])
-        #     and _check_shape(bindings, value_mm_reshaped, ["B", "S", "H", "d_h"])
-        #     and _check_shape(bindings, key_reshaped, ["B*H", "KVS", "d_h"])
-        #     and _check_shape(bindings, key_transposed, ["B", "H", "d_h", "KVS"])
-        #     and _check_shape(bindings, attention_reshaped, ["B", "S", "H*d_h"])
-        # )
-        # if not status:
+        # bindings: dict[str, Dim] = {}
+
+        # def no_match(val: ir.Value, dims: Sequence[str]) -> bool:
+        #     return not _check_shape(bindings, val, dims)
+
+        # if no_match(query_BSD, ["B", "S", "D"]):
         #     return False
-        # if bindings["B"] * bindings["H"] != bindings["B*H"]:
+        # if no_match(key_BSDkv, ["B", "Skv", "D"]):
         #     return False
-        # if bindings["H"] * bindings["d_h"] != bindings["H*d_h"]:
+        # if no_match(value_BSDkv, ["B", "Skv", "D"]):
         #     return False
+
+        # if no_match(past_key, ["B", "H", "Spast", "Dh"]):
+        #     return False
+        # if no_match(past_value, ["B", "H", "Spast", "Dv"]):
+        #     return False
+        # if no_match(query_BSHDh, ["B", "S", "H", "Dh"]):
+        #     return False
+        # if no_match(key_BSHkvDh, ["B", "S", "H", "Dh"]):
+        #     return False
+        # if no_match(value_BSHkvDh, ["B", "S", "H", "Dh"]):
+        #     return False
+
+        # TODO: mask shape check: ideally, it should be (1 or B, 1 or H, S, St)
+        # But this also, unforunately, depends on ORT version.
+
+        # TODO: verify Reshapes:
+        # eg.: verify bindings["B"] * bindings["H"] == bindings["B*H"]:
+        # and bindings["H"] * bindings["Dh"] == bindings["H*Dh"]:
+        # or check Reshape's shape-input value
+
         return True
 
     def rewrite(
         self,
         op,
-        input,
-        qkv_weight,
+        query_BSD,
+        key_BSDkv,
+        value_BSDkv,
         mask,
-        cos,
-        sin,
         past_key,
         past_value,
-        position_ids,
-        query_mm_reshaped,
+        # key_BSHkvDh,
+        # position_ids,
+        # cos,
+        # sin,
         **_,
     ):
-        num_heads = query_mm_reshaped.shape[2]
-        qkv = op.MatMul(input, qkv_weight)
-        return op.GroupQueryAttention(
-            qkv,
-            None,  # key
-            None,  # value
+        # num_heads = _ir_utils.get_dim(key_BSHkvDh, 2)
+        # if not isinstance(num_heads, int):
+        #     return None
+
+        # # Switch to 3D RotaryEmbedding
+        # # TODO: forward other attributes
+        # query_BSD_rope = op.RotaryEmbedding(
+        #     query_BSD, position_ids, cos, sin, _domain="com.microsoft"
+        # )
+        # key_BSD_rope = op.RotaryEmbedding(
+        #     key_BSDkv, position_ids, cos, sin, _domain="com.microsoft"
+        # )
+
+        return op.DummyGQA(
+            query_BSD,
+            key_BSDkv,
+            value_BSDkv,
+            None,  # bias
+            None,  # key padding mask
+            mask,  # attention mask/bias
             past_key,
             past_value,
-            # seqlens_k,
-            # total_sequence_length,
-            cos,
-            sin,
-            num_heads=num_heads,
+            # num_heads=num_heads,
             _domain="com.microsoft",
             _outputs=3,
         )
 
 
-_rule1 = GroupQueryAttention.rule("MHA_2dmm", use_2d_matmul=False)
+_rule1 = GroupQueryAttention.rule()
 
 gqa_rules = pattern.RewriteRuleSet([_rule1])
 
@@ -152,5 +235,5 @@ gqa_rules = pattern.RewriteRuleSet([_rule1])
 def fuse_gqa(model: ir.Model) -> int:
     count = gqa_rules.apply_to_model(model)
     print(f"GQA count: {count}")
-    remove_unused_nodes(model)
+    # remove_unused_nodes(model)
     return count
